💄 Improve job page heading
